requirements.txt
aniso8601==7.0.0
asgiref==3.2.3
astroid==2.3.3
autopep8==1.4.4
Click==7.0
cmake==3.15.3
colorama==0.4.3
Django==3.0
django-cors-headers==3.2.0
django-graphql-jwt==0.3.0
dlib==19.15.0
face-recognition==1.2.3
face-recognition-models==0.3.0
graphene==2.1.8
graphene-django==2.7.1
graphene-file-upload==1.2.2
graphql-core==2.2.1
graphql-relay==2.0.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.17.4
Pillow==6.2.1
promise==2.2.1
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.4.4
pytz==2019.3
Rx==1.6.1
singledispatch==3.4.0.3
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.2
